Added pre-defined vinyl state data, change model, new album request, simpify image service

+++ Web/Models/Album.cs
@@ -18,7 +18,7 @@
         /// <summary>
         /// Album size, Gb
         /// </summary>
-        public float? Size { get; set; }
+        public double? Size { get; set; }
 
         /// <summary>
         /// Source from where album was downloaded / buyed
